vegetationFLOW_tool/requirements.txt
fastapi
uvicorn
celery
redis
streamlit
folium
streamlit_folium
pyyaml